# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_workbook_api_workspaces__workspace_id__replica_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_workbook_api_workspaces__workspace_id__replica_import_post",
        "type": "object"
      },
...
      "Column": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "predicate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicate"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "unit_y": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Y"
          }
        },
        "required": [
          "name"
        ],
        "title": "Column",
        "type": "object"
      },
      "Component": {
        "properties": {
          "cls": {
            "title": "Cls",
            "type": "string"
          },
          "columns": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "title": "Columns",
            "type": "array"
          },
          "rows": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "cls"
        ],
        "title": "Component",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplicaSpec": {
        "properties": {
          "components": {
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "title": "Components",
            "type": "array"
          },
          "persist": {
            "default": true,
            "title": "Persist",
            "type": "boolean"
          }
        },
        "required": [
          "components"
        ],
        "title": "ReplicaSpec",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP access to the Digicities platform backend: the same functions the Streamlit app calls in-process, exposed for the React frontend (`digicities-frontend`) and any other client.\n\nAlmost every route is scoped to a workspace via `/api/workspaces/{workspace_id}`; start with **workspaces** to find or create one.\n\nAuthentication is off by default (open local mode). With `API_AUTH_ENABLED` set, every route requires an `Authorization: Bearer <token>` header (Keycloak token; see `apps/api/auth.py`).",
    "title": "Digicities API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspaces/{workspace_id}/replica/config": {
      "get": {
        "operationId": "config_api_workspaces__workspace_id__replica_config_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Config Api Workspaces  Workspace Id  Replica Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Config",
        "tags": [
          "replica"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/replica/generate": {
      "post": {
        "description": "Build a workbook from the in-app replica model and convert it to instance TTL.",
        "operationId": "generate_api_workspaces__workspace_id__replica_generate_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplicaSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Api Workspaces  Workspace Id  Replica Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "replica"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/replica/import": {
      "post": {
        "description": "Convert an uploaded digital-replica workbook (.xlsx) to instance TTL via\n``process_excel_to_ttl`` and persist it to the workspace's ingestion output.",
        "operationId": "import_workbook_api_workspaces__workspace_id__replica_import_post",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_workbook_api_workspaces__workspace_id__replica_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Workbook Api Workspaces  Workspace Id  Replica Import Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Workbook",
        "tags": [
          "replica"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/replica/model": {
      "get": {
        "description": "The workspace's current replica TTL parsed back into the in-app model —\nthe round-trip enabler for an in-app editor.\n\nResponse shape::\n\n    {\n      \"file\": \"<workspace>.ttl\" | null,\n      \"instances\": [{id, component_type, uri, label,\n                     attributes: {name: {type, ...}},\n                     annotations: {...}, class_objects: {...}}, ...],\n      \"draft\": {\"components\": [{cls, columns: [{name, type, unit, unit_y,\n                                                currency, predicate, key}],\n                                rows: [...]}]}\n    }\n\n``draft`` is the same :class:`ReplicaDraft` schema ``POST /generate``\naccepts, so a client can GET the model, edit it, and POST it back.",
        "operationId": "replica_model_api_workspaces__workspace_id__replica_model_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replica Model Api Workspaces  Workspace Id  Replica Model Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replica Model",
        "tags": [
          "replica"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/replica/ttl": {
      "get": {
        "description": "The workspace's current replica TTL (Preview & Export).",
        "operationId": "replica_ttl_api_workspaces__workspace_id__replica_ttl_get",
        "parameters": [
          {
            "description": "workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "workspace id",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replica Ttl Api Workspaces  Workspace Id  Replica Ttl Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replica Ttl",
        "tags": [
          "replica"
        ]
      }
    },
...
  },
...
}